fix: allow rendering form helpers with blocks
This change implements activeadmin/arbre#64 https://github.com/activeadmin/arbre/pull/64/files which was reverted.

The initial PR was reverted because you could potentially get duplicate renders in certain situations. The workaround I can see is to return nil from the block if this occurs. I think this is an acceptable workaround for this library.